+++ CCHMapClusterController Example iOS/CCHMapClusterController Example iOS/ClusterAnnotationView.m
@@ -10,28 +10,6 @@
 
 #import "ClusterAnnotationView.h"
 
-static inline CGPoint TBRectCenter(CGRect rect)
-{
-    return CGPointMake(CGRectGetMidX(rect), CGRectGetMidY(rect));
-}
-
-static inline CGRect TBCenterRectRounded(CGRect rect, CGPoint center)
-{
-    CGRect r = CGRectMake(roundf(center.x - rect.size.width/2.0),
-                          roundf(center.y - rect.size.height/2.0),
-                          roundf(rect.size.width),
-                          roundf(rect.size.height));
-    return r;
-}
-
-static CGFloat const TBScaleFactorAlpha = 0.3;
-static CGFloat const TBScaleFactorBeta = 0.4;
-
-static inline CGFloat TBScaledValueForValue(CGFloat value)
-{
-    return 1.0 / (1.0 + expf(-1 * TBScaleFactorAlpha * powf(value, TBScaleFactorBeta)));
-}
-
 @interface ClusterAnnotationView ()
 
 @property (strong, nonatomic) UILabel *countLabel;
@@ -45,21 +23,22 @@ static inline CGFloat TBScaledValueForValue(CGFloat value)
     self = [super initWithAnnotation:annotation reuseIdentifier:reuseIdentifier];
     if (self) {
         self.backgroundColor = [UIColor clearColor];
-        [self setupLabel];
+        [self setUpLabel];
         [self setCount:1];
     }
     return self;
 }
 
-- (void)setupLabel
+- (void)setUpLabel
 {
-    _countLabel = [[UILabel alloc] initWithFrame:self.frame];
+    _countLabel = [[UILabel alloc] initWithFrame:self.bounds];
     _countLabel.autoresizingMask = UIViewAutoresizingFlexibleWidth | UIViewAutoresizingFlexibleHeight;
     _countLabel.textAlignment = NSTextAlignmentCenter;
     _countLabel.backgroundColor = [UIColor clearColor];
     _countLabel.textColor = [UIColor whiteColor];
     _countLabel.textAlignment = NSTextAlignmentCenter;
     _countLabel.adjustsFontSizeToFitWidth = YES;
+    _countLabel.minimumScaleFactor = 2;
     _countLabel.numberOfLines = 1;
     _countLabel.font = [UIFont boldSystemFontOfSize:12];
     _countLabel.baselineAdjustment = UIBaselineAdjustmentAlignCenters;
@@ -71,38 +50,36 @@ static inline CGFloat TBScaledValueForValue(CGFloat value)
 {
     _count = count;
     
-    CGPoint oldCenter = self.center;
-    CGRect newBounds = CGRectMake(0, 0, 44 * TBScaledValueForValue(count), 44 * TBScaledValueForValue(count));
-    self.frame = TBCenterRectRounded(newBounds, oldCenter);
-    self.center = oldCenter;
-
     self.countLabel.text = [@(count) stringValue];
-    
-    [self setNeedsDisplay];
+    [self update];
 }
 
-- (void)drawRect:(CGRect)rect
+- (void)update
 {
-    CGContextRef context = UIGraphicsGetCurrentContext();
+    self.countLabel.frame = self.bounds;
     
-    CGContextSetAllowsAntialiasing(context, true);
-    
-    UIColor *outerCircleStrokeColor = [UIColor colorWithWhite:0 alpha:0.25];
-    UIColor *innerCircleStrokeColor = [UIColor whiteColor];
-    UIColor *innerCircleFillColor = [UIColor colorWithRed:(255.0 / 255.0) green:(95 / 255.0) blue:(42 / 255.0) alpha:1.0];
-    
-    CGRect circleFrame = CGRectInset(rect, 4, 4);
-    
-    [outerCircleStrokeColor setStroke];
-    CGContextSetLineWidth(context, 5.0);
-    CGContextStrokeEllipseInRect(context, circleFrame);
-    
-    [innerCircleStrokeColor setStroke];
-    CGContextSetLineWidth(context, 4);
-    CGContextStrokeEllipseInRect(context, circleFrame);
-    
-    [innerCircleFillColor setFill];
-    CGContextFillEllipseInRect(context, circleFrame);
+    // Images are faster than using drawRect:
+    UIImage *image;
+    if (self.count > 1000) {
+        image = [UIImage imageNamed:@"Circle39"];
+    } else if (self.count > 500) {
+        image = [UIImage imageNamed:@"Circle38"];
+    } else if (self.count > 200) {
+        image = [UIImage imageNamed:@"Circle36"];
+    } else if (self.count > 100) {
+        image = [UIImage imageNamed:@"Circle34"];
+    } else if (self.count > 50) {
+        image = [UIImage imageNamed:@"Circle31"];
+    } else if (self.count > 20) {
+        image = [UIImage imageNamed:@"Circle28"];
+    } else if (self.count > 10) {
+        image = [UIImage imageNamed:@"Circle25"];
+    } else if (self.count > 5) {
+        image = [UIImage imageNamed:@"Circle24"];
+    } else {
+        image = [UIImage imageNamed:@"Circle21"];
+    }
+    self.image = image;
 }
 
 @end